# core-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Make-Singer — AI-powered music creation platform.",
        "title": "Make-Singer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/export/mp3": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export the mastered track as an MP3 file with optional metadata",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Export as MP3",
                "parameters": [
                    {
                        "description": "MP3 export request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ExportMP3Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExportMP3Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/export/stems": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export individual stems as a bundled archive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Export stems",
                "parameters": [
                    {
                        "description": "Stems export request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ExportStemsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExportStemsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/export/wav": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Export the mastered track as a WAV file with configurable bit depth and sample rate",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Export"
                ],
                "summary": "Export as WAV",
                "parameters": [
                    {
                        "description": "WAV export request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ExportWAVRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ExportWAVResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lyrics/generate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generate lyrics for a song section using AI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lyrics"
                ],
                "summary": "Generate lyrics",
                "parameters": [
                    {
                        "description": "Generate request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LyricsGenerateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LyricsGenerateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/lyrics/rewrite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Rewrite existing lyrics with AI-powered suggestions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lyrics"
                ],
                "summary": "Rewrite lyrics",
                "parameters": [
                    {
                        "description": "Rewrite request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LyricsRewriteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.LyricsRewriteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.ExportMP3Request": {
            "type": "object",
            "required": [
                "masterFileUrl",
                "projectId"
            ],
            "properties": {
                "masterFileUrl": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/model.ExportMetadata"
                },
                "projectId": {
                    "type": "string"
                },
                "quality": {
                    "type": "integer",
                    "enum": [
                        128,
                        192,
                        256,
                        320
                    ]
                }
            }
        },
        "model.ExportMP3Response": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "quality": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "model.ExportMetadata": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "string",
                    "maxLength": 200
                },
                "artist": {
                    "type": "string",
                    "maxLength": 200
                },
                "credits": {
                    "type": "string",
                    "maxLength": 1000
                },
                "title": {
                    "type": "string",
                    "maxLength": 200
                },
                "year": {
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 1900
                }
            }
        },
        "model.ExportStemsRequest": {
            "type": "object",
            "required": [
                "projectId",
                "stemUrls"
            ],
            "properties": {
                "format": {
                    "type": "string",
                    "enum": [
                        "wav",
                        "mp3"
                    ]
                },
                "includeMaster": {
                    "type": "boolean"
                },
                "includeVocals": {
                    "type": "boolean"
                },
                "masterUrl": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "stemUrls": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "vocalUrls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.ExportStemsResponse": {
            "type": "object",
            "properties": {
                "expiresAt": {
                    "type": "string"
                },
                "fileCount": {
                    "type": "integer"
                },
                "fileUrl": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "model.ExportWAVRequest": {
            "type": "object",
            "required": [
                "masterFileUrl",
                "projectId"
            ],
            "properties": {
                "bitDepth": {
                    "type": "integer",
                    "enum": [
                        16,
                        24,
                        32
                    ]
                },
                "masterFileUrl": {
                    "type": "string"
                },
                "projectId": {
                    "type": "string"
                },
                "sampleRate": {
                    "type": "integer",
                    "enum": [
                        44100,
                        48000,
                        96000
                    ]
                }
            }
        },
        "model.ExportWAVResponse": {
            "type": "object",
            "properties": {
                "bitDepth": {
                    "type": "integer"
                },
                "expiresAt": {
                    "type": "string"
                },
                "fileUrl": {
                    "type": "string"
                },
                "format": {
                    "type": "string"
                },
                "sampleRate": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "model.Genre": {
            "type": "string",
            "enum": [
                "pop",
                "rock",
                "hiphop",
                "rnb",
                "electronic",
                "jazz",
                "country",
                "folk",
                "classical",
                "latin",
                "reggae",
                "blues"
            ],
            "x-enum-varnames": [
                "GenrePop",
                "GenreRock",
                "GenreHiphop",
                "GenreRnb",
                "GenreElectronic",
                "GenreJazz",
                "GenreCountry",
                "GenreFolk",
                "GenreClassical",
                "GenreLatin",
                "GenreReggae",
                "GenreBlues"
            ]
        },
...
        "model.Language": {
            "type": "string",
            "enum": [
                "en",
                "tr",
                "fr"
            ],
            "x-enum-varnames": [
                "LanguageEN",
                "LanguageTR",
                "LanguageFR"
            ]
        },
        "model.LyricsGenerateRequest": {
            "type": "object",
            "required": [
                "genre",
                "sectionType",
                "vibes"
            ],
            "properties": {
                "genre": {
                    "enum": [
                        "pop",
                        "rock",
                        "hiphop",
                        "rnb",
                        "electronic",
                        "jazz",
                        "country",
                        "folk",
                        "classical",
                        "latin",
                        "reggae",
                        "blues"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Genre"
                        }
                    ]
                },
                "language": {
                    "enum": [
                        "en",
                        "tr",
                        "fr"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Language"
                        }
                    ]
                },
                "sectionType": {
                    "enum": [
                        "intro",
                        "verse",
                        "prechorus",
                        "chorus",
                        "bridge",
                        "outro",
                        "instrumental"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SectionType"
                        }
                    ]
                },
                "vibes": {
                    "type": "array",
                    "maxItems": 3,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.LyricsGenerateResponse": {
            "type": "object",
            "properties": {
                "drafts": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "model.LyricsRewriteRequest": {
            "type": "object",
            "required": [
                "currentLyrics",
                "genre",
                "sectionType",
                "vibes"
            ],
            "properties": {
                "currentLyrics": {
                    "type": "string",
                    "minLength": 1
                },
                "genre": {
                    "enum": [
                        "pop",
                        "rock",
                        "hiphop",
                        "rnb",
                        "electronic",
                        "jazz",
                        "country",
                        "folk",
                        "classical",
                        "latin",
                        "reggae",
                        "blues"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Genre"
                        }
                    ]
                },
                "instructions": {
                    "type": "string",
                    "maxLength": 500
                },
                "sectionType": {
                    "enum": [
                        "intro",
                        "verse",
                        "prechorus",
                        "chorus",
                        "bridge",
                        "outro",
                        "instrumental"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SectionType"
                        }
                    ]
                },
                "vibes": {
                    "type": "array",
                    "maxItems": 3,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.LyricsRewriteResponse": {
            "type": "object",
            "properties": {
                "lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.SectionType": {
            "type": "string",
            "enum": [
                "intro",
                "verse",
                "prechorus",
                "chorus",
                "bridge",
                "outro",
                "instrumental"
            ],
            "x-enum-varnames": [
                "SectionIntro",
                "SectionVerse",
                "SectionPrechorus",
                "SectionChorus",
                "SectionBridge",
                "SectionOutro",
                "SectionInstrumental"
            ]
        },
...
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}